# [FIX] stock_sale: imp lot._compute_sale_order_ids perf.