Improve _compute_sale_order_ids performances by saving
stock.move.line search results to defaultdict before going
through records in self.


##### speedup

DB: 1939 lots, 110k stock_moves, 100k move_lines
Reported Average Time of
```python 
lots.modified(fnames=['name'])
lots.recompute()
```

| len(lots) | before PR | after PR |
|:--------------------:|:-------------:|:------------:|
| 50 | 1.61s | 0.05s |
| 200 | 5s | 0.09s |
| 500 | 11.29s | 0.18s |
| 1000 | 21.46s | 0.36s |
| 1939 | 40.7s | 0.69s |

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr